Move initOffsets functionality to a common package
